+++ docs/assets/js/096bfee4.dc52abac.js
@@ -1,1 +1,0 @@
-"use strict";(self.webpackChunkrxdb=self.webpackChunkrxdb||[]).push([[7178],{5010:e=>{e.exports=JSON.parse('{"permalink":"/blog/tags/facebook","page":1,"postsPerPage":10,"totalPages":1,"totalCount":1,"blogDescription":"Blog","blogTitle":"Blog"}')}}]);