+++ docs/assets/js/30a24c52.2abbf3d8.js
@@ -1,1 +1,0 @@
-"use strict";(self.webpackChunkrxdb=self.webpackChunkrxdb||[]).push([[453],{8605:l=>{l.exports=JSON.parse('{"label":"hello","permalink":"/blog/tags/hello","allTagsPath":"/blog/tags","count":2,"unlisted":false}')}}]);